+++ Cargo.toml
@@ -118,6 +118,7 @@ memory-stats = { version = "1", default-features = false }
 simdutf8 = { version = "0.1.4", default-features = false }
 smallvec = { version = "1.13", default-features = false }
 windows-sys = { version = "0.52", default-features = false }
+cgroupfs = { version = "0.8", default-features = false }
 
 [patch.crates-io]
 # Git dependency for `containerd-client` to:
