+++ lib/saluki-env/src/workload/collectors/mod.rs
@@ -4,7 +4,8 @@ use tracing::{debug, error};
 
 use super::metadata::MetadataOperation;
 
-mod cgroupsv2;
+mod cgroups_v2;
+pub use self::cgroups_v2::CGroupsV2MetadataCollector;
 
 mod containerd;
 pub use self::containerd::ContainerdMetadataCollector;
